+++ CHANGELOG.md
@@ -2,8 +2,13 @@
 
 ## v1.1.0
 - [Add layers attribute](https://github.com/babbel/terraform-aws-lambda-with-inline-code/pull/7)
+- [Add secret environment variables attribute](https://github.com/babbel/terraform-aws-lambda-with-inline-code/pull/8)
 
 
 ## v1.0.0
 
 - [Initial version](https://github.com/babbel/terraform-aws-lambda-with-inline-code/pull/1)
+
+## v1.1.0
+
+- [Added support for secret environment variables](https://github.com/babbel/terraform-aws-lambda-with-inline-code/pull/6)